+++ auto_route/lib/annotations.dart
@@ -1,1 +1,2 @@
 export 'src/common/auto_route_annotations.dart';
+export 'src/common/param_converter.dart';